+++ SimTKmath/Geometry/src/ContactGeometryImpl.h
@@ -140,7 +140,7 @@ public:
     bool calcNearestPointOnLineImplicitly(
             const Vec3& pointA,
             const Vec3& pointB,
-            size_t maxIterations,
+            int maxIterations,
             Real tolerance,
             Vec3& nearestPointOnLine) const;
